config: Apply output config also during config reload.
